chore(log): moved commit and git log entry types to types dir
