+++ packages/library/src/GitLog.integration.spec.tsx
@@ -9,7 +9,7 @@ import { GitLog } from './GitLog'
 import { sleepCommits } from 'test/data/sleepCommits'
 import { commitNode } from 'test/elements/CommitNode'
 import { tag } from 'test/elements/Tag'
-import { Commit } from 'types'
+import { Commit } from 'types/Commit'
 import { formatBranch } from 'modules/Tags/utils/formatBranch'
 import { table } from 'test/elements/Table'
 import dayjs from 'dayjs'
