+++ packages/library/src/_test/data/sleepCommits.ts
@@ -1,4 +1,4 @@
-import { Commit } from 'types'
+import { Commit } from 'types/Commit'
 
 export const sleepCommits: Commit[] = [
   {
