+++ packages/library/src/hooks/useTheme/types.ts
@@ -1,4 +1,4 @@
-import { Commit } from 'types'
+import { Commit } from 'types/Commit'
 
 export interface ThemeFunctions {
   theme: ThemeMode
