+++ packages/library/src/modules/Graph/components/GraphColumn/types.ts
@@ -1,4 +1,4 @@
-import { Commit } from 'types'
+import { Commit } from 'types/Commit'
 
 export interface GraphColumnProps {
   /**
