+++ packages/library/src/modules/Graph/hooks/usePlaceholderData/types.ts
@@ -1,4 +1,4 @@
-import { Commit } from 'types'
+import { Commit } from 'types/Commit'
 import { GraphColumnState } from 'modules/Graph/components/GraphColumn'
 
 export interface PlaceholderDatum {
