+++ packages/library/src/modules/Table/components/TableRow/types.ts
@@ -1,4 +1,4 @@
-import { Commit } from 'types'
+import { Commit } from 'types/Commit'
 
 export interface GitLogTableRowProps {
   index: number
