+++ packages/library/src/modules/Tags/components/BranchIcon/BranchIcon.tsx
@@ -1,3 +1,4 @@
+// eslint-disable-next-line import/extensions
 import Icon from 'assets/branch.svg?react'
 import { useTheme } from 'hooks/useTheme'
 import styles from './BranchIcon.module.scss'
