+++ packages/library/src/modules/Tags/components/BranchTag/types.ts
@@ -1,4 +1,4 @@
-import { Commit } from 'types'
+import { Commit } from 'types/Commit'
 
 export interface BranchTagProps {
   id: string
