+++ packages/library/src/modules/Tags/components/BranchTagTooltip/types.ts
@@ -1,4 +1,4 @@
-import { Commit } from 'types'
+import { Commit } from 'types/Commit'
 
 export interface BranchTagTooltipProps {
   id: string
