+++ packages/library/src/modules/Tags/components/TagIcon/TagIcon.tsx
@@ -1,3 +1,4 @@
+// eslint-disable-next-line import/extensions
 import Icon from 'assets/tag.svg?react'
 import { useTheme } from 'hooks/useTheme'
 import styles from './TagIcon.module.scss'
