+++ packages/library/src/types.ts
@@ -1,5 +1,7 @@
 import { CSSProperties } from 'react'
 import { ThemeColours, ThemeMode } from './hooks/useTheme/types'
+import { GitLogEntry } from './types/GitLogEntry'
+import { Commit } from './types/Commit'
 
 export interface GitLogProps {
   /**
@@ -155,129 +157,4 @@ export interface GitLogPaging {
    * The first page is page 0.
    */
   page: number
-}
-
-/**
- * Represents a commit in the Git history.
- */
-export interface Commit {
-  /**
-   * The unique hash (SHA) identifying the commit.
-   */
-  hash: string
-
-  /**
-   * An array of parent commit hashes (SHA) for this commit.
-   * A commit can have multiple parents in the case of merges.
-   */
-  parents: string[]
-
-  /**
-   * An array of child commit hashes (SHA) that
-   * reference this commit as a parent.
-   *
-   * This helps track descendants in the commit graph.
-   */
-  children: string[]
-
-  /**
-   * The name of the branch this commit belongs to.
-   */
-  branch: string
-
-  /**
-   * The commit message describing the changes
-   * introduced by this commit.
-   */
-  message: string
-
-  /**
-   * Details of the user who authored
-   * the commit.
-   */
-  author?: CommitAuthor
-
-  /**
-   * The date and time when the commit was
-   * made by the author, in ISO 8601 format.
-   */
-  authorDate?: string
-
-  /**
-   * The date and time when the commit was
-   * committed to the repository, in ISO 8601 format.
-   *
-   * This may differ from `authorDate` in cases
-   * like rebases or amend commits.
-   */
-  committerDate: string
-
-  /**
-   * Indicates whether this commit is the
-   * tip (latest commit) of its branch.
-   */
-  isBranchTip: boolean
-}
-
-/**
- * Represents the author of a Git commit.
- */
-export interface CommitAuthor {
-  /**
-   * The name of the commit author.
-   */
-  name?: string;
-
-  /**
-   * The email address of the commit author.
-   */
-  email?: string;
-}
-
-/**
- * Represents a single entry in the git log.
- */
-export interface GitLogEntry {
-  /**
-   * The unique hash identifier of the commit.
-   */
-  hash: string
-
-  /**
-   * The name of the branch this commit belongs to.
-   */
-  branch: string
-
-  /**
-   * An array of parent commit hashes.
-   *
-   * - If this commit is a merge commit, it will have multiple parents.
-   * - If this commit is an initial commit, it will have no parents.
-   */
-  parents: string[]
-
-  /**
-   * The commit message describing the changes made in this commit.
-   */
-  message: string
-
-  /**
-   * Details of the user who authored
-   * the commit.
-   */
-  author?: CommitAuthor
-
-  /**
-   * The date and time when the commit was applied by the committer.
-   *
-   * This is typically the timestamp when the commit was finalized.
-   */
-  committerDate: string
-
-  /**
-   * The date and time when the commit was originally authored.
-   *
-   * This may differ from `committerDate` if the commit was rebased or amended.
-   */
-  authorDate?: string
 }
